+++ bindings/python/parsers/mjcf/model.cpp
@@ -1,5 +1,5 @@
 //
-// Copyright (c) 2024 CNRS INRIA
+// Copyright (c) 2024 INRIA
 //
 
 #include "pinocchio/parsers/mjcf.hpp"
@@ -34,14 +34,14 @@ namespace pinocchio
         "buildModelFromMJCF",
         static_cast<Model (*)(const std::string &)>(pinocchio::python::buildModelFromMJCF),
         bp::args("mjcf_filename"),
-        "Parse the Mjcf file given in input and return a pinocchio Model");
+        "Parse the MJCF file given in input and return a pinocchio Model.");
 
       bp::def(
         "buildModelFromMJCF",
         static_cast<Model (*)(const std::string &, const JointModel &)>(
           pinocchio::python::buildModelFromMJCF),
         bp::args("mjcf_filename", "root_joint"),
-        "Parse the Mjcf file and return a pinocchio Model with the given root Joint.");
+        "Parse the MJCF file and return a pinocchio Model with the given root Joint.");
     }
   } // namespace python
 } // namespace pinocchio